Refactor to remove duplicate code, fixing a couple of bugs
When sending a POST, the Content-Type was missing. This caused problems
with some services (e.g. Twitter).
When there was only one parameter it was encoded as param=value&
(with a extra & at the end). I didn't find any service which rejected
this invalid encoding, but is better to fix it just in case.